feature: added logged out endpoint for web
