+++ src/handlers/api/auth/authenticated/revoke_token.rs
@@ -3,9 +3,9 @@ use std::sync::Arc;
 use axum::{extract::State, response::IntoResponse};
 use tower_cookies::Cookies;
 
-use crate::model::{error, AppState};
-use crate::middleware::{auth::{self, Ctx}, cookies::Manager};
-use crate::server_error;
+use crate::handlers::logic;
+use crate::model::AppState;
+use crate::middleware::auth::Ctx;
 
 //can see this as a logout
 pub async fn revoke_token(
@@ -14,22 +14,5 @@ pub async fn revoke_token(
     jar: Cookies,
 ) -> impl IntoResponse
 {
-    let repo_refresh = &state.refresh_tokens;
-
-    let device_id_cookie = jar.get_cookie(auth::CookieNames::DEVICE_ID.as_str())
-        .map_err(|err| server_error!(err, error::Kind::NoAuth, error::OnType::Cookie))?;
-    
-    let ctx_user_id = &ctx.user_id_ref();
-
-    match repo_refresh.revoke_token(ctx_user_id, &device_id_cookie).await
-    {
-        Ok(()) => 
-        {
-            jar.remove_cookie(auth::CookieNames::AUTH_ACCES.to_string());
-            jar.remove_cookie(auth::CookieNames::AUTH_REFRESH.to_string());
-
-            Ok(())
-        },
-        Err(err) => Err(err),
-    }
+    logic::auth::authenticated::revoke_token(&state, &ctx, &jar).await
 }
