+++ src/handlers/logic/auth/authenticated.rs
@@ -1,0 +1,3 @@
+mod revoke_token;
+
+pub use revoke_token::*;
