+++ src/handlers/web/auth/authenticate.rs
@@ -1,0 +1,3 @@
+mod logout;
+
+pub use logout::*;
